Reorganize codebase into models, config, and utils.py, begin FastAPI implementation
