+++ .gitignore
@@ -2,6 +2,5 @@
 __pycache__
 .venv/
 notes.txt
-emailtest.py
 stock_data_polygon.py
 main310.py
