+++ requirements.txt
@@ -14,3 +14,4 @@ sniffio==1.3.1
 typing_extensions==4.12.2
 SQLAlchemy==2.0.36
 psycopg2==2.9.10
+fastapi==0.115.4
